add base eval
